Persist address associations properly
The `AssociateAddress` API is currently implemented simply by running
`ip addr add` on the domain VM.  This is not durable as the IP
configuration will be blown away by the next networkd reload which might
happen for any reason.

Fix this by placing the EIP address in a `.network.d` drop-in instead
making it survive networkd reloads and even VM reboots.